requirements.txt
asgiref==3.7.2
Django==5.0.1
django-ckeditor==6.7.0
django-js-asset==2.2.0
djangorestframework==3.14.0
pillow==10.2.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
